Add a OurQuotes plugin to accept and recall quotes and quips
I want to import a collection of quotes from an IRC Eggdrop plugin
and make it accessible over Campfire.

Store changes to the quote archive like the Eggdrop version, and
rebuild an in-memory summary of the log each time a change is made.

Allow adding new quotes with !addquote, removing existing quotes with
!rmquote, and recalling quotes with !quote <id>, !quote <searchstr>,
and !quote.

Set the default quote recall command to be '!ourquote' instead of
'!quote' so it can work out-of-the-box and still avoid clobbering the
previously existing quote plugin.